Find unique
here,  the concept of XOR has been used.
as it is mentioned that "m" numbers of element are repeated and there is only one which occur once. so for finding that number if we  XOR all the elements (for the same no. XOR is 0), then all the repeated one will become 0 and only that one number will remain.